Added checking for value of constants.

+++ src/Constant.cpp
@@ -8,7 +8,7 @@ using std::string;
 using std::to_string;
 using std::shared_ptr;
 
-Constant::Constant(const int& value, const shared_ptr<Type>& type) {
+Constant::Constant(const long long int &value, const shared_ptr<Type> &type) {
     this->value = value;
     this->type = type;
 }
@@ -17,11 +17,11 @@ Constant::~Constant() {
 
 }
 
-void Constant::setValue(const int &value) {
+void Constant::setValue(const long long int &value) {
     this->value = value;
 }
 
-int Constant::getValue() {
+long long int Constant::getValue() {
     return value;
 }
